fix: resolve system-acceleration compilation errors
- Fix Core ML bindings and function signatures
- Add missing tokenizer functions for Mistral model
- Unify tensor types with candle_core::Tensor
- Fix import paths and type references
- Add missing Core ML stub implementations
- Resolve ANEError type conversions

All compilation errors in system-acceleration crate resolved.

+++ iterations/v3/system-acceleration/src/ane/infer/whisper.rs
@@ -221,7 +221,7 @@ impl WhisperInferenceExecutor {
             // Reshape mel spectrogram for CoreML input
             // Whisper encoder expects [1, 80, 3000] shape
             let mel_data = &input.mel_spectrogram;
-            let input_shape = [1i32, input.n_mels as i32, input.n_time_steps as i32];
+            let input_shape = [1usize, input.n_mels as usize, input.n_time_steps as usize];
 
             // Run inference on the encoder
             let _output_tensor = coreml::run_inference(
